Fix exclusive service tests for PRs
