+++ package.json
@@ -45,6 +45,9 @@
     "lint": "eslint '**/*.js'",
     "test:js": "mocha lib '*.spec.js'",
     "test:services": "mocha --delay service-tests/runner/cli.js",
+    "test:services:pr:prepare": "node service-tests/runner/pull-request-services-cli.js > pull-request-services.log",
+    "test:services:pr:run": "mocha --delay service-tests/runner/cli.js --stdin < pull-request-services.log",
+    "test:services:pr": "npm run test:services:pr:prepare && npm run test:services:pr:run",
     "test": "npm run lint && npm run test:js"
   },
   "bin": {
@@ -76,6 +79,7 @@
     "nock": "^9.0.13",
     "node-fetch": "^1.6.3",
     "opn-cli": "^3.1.0",
+    "read-all-stdin-sync": "^1.0.5",
     "sinon": "^2.1.0"
   },
   "engines": {
